backend/requirements/local.txt
# =========================
# ① Django 本体・依存ライブラリ
# =========================
Django>=5.2.2            # Django 本体（Webアプリのフレームワーク）
asgiref                  # Django が async/ASGI を扱うための内部依存
sqlparse                 # Django が SQL を扱うための内部依存

# =========================
# ② 環境変数・設定管理
# =========================
django-environ          # .env の読み込み、Django 設定を環境変数で管理しやすくする
python-dotenv
dj_database_url         # DATABASE_URL を Django 設定形式に変換（Render・Railway 用）
datetime

# =========================
# ③ REST API・認証関連
# =========================

djangorestframework>=3.16.0   # REST API を作るためのフレームワーク
djangorestframework-simplejwt # JWT 認証（フロントと連携しやすい）
dj-rest-auth==6.0.0            # ログイン / ログアウト / パスワードリセット等 API を自動生成
django-allauth>=65.9.0 # SNSログインやメール認証。socialaccount使わないなら不要
requests                        # 外部APIへの HTTP リクエスト
django-timezone-field           # TimeZoneField をモデルで使う場合に必要
drf-nested-routers              # ネストされたルーティングを簡単に定義するためのライブラリ

# =========================
# ④ 本番デプロイ関連
# =========================
gunicorn                # Linuxサーバーで Django を動かすアプリサーバー
psycopg2-binary         # PostgreSQL 用 DB ドライバ
redis                   # Redis クライアント（キャッシュやキュー）

# =========================
# ⑤ コメントアウト済み / 使ってない可能性あり
# =========================
# celery                 # 非同期タスクキュー（使ってないなら不要）
# django-celery-beat     # Celery の定期実行（Celery未使用なら不要）
# django-cors-headers    # CORS 制御用（INSTALLED_APPS に無ければ使えてない）
# django-hide            # 管理画面でフィールド非表示。使ってないなら不要
